add android test
