+++ ext/src/commonMain/kotlin/top/kagg886/mkmb/ext/center.kt
@@ -127,4 +127,4 @@ fun MMKV.addEventListener(listener: (String) -> Unit): () -> Unit {
     }
 }
 
-expect fun <T: Any> getFromPlatform(clazz: KClass<T>,buffer: Buffer): T
+internal expect fun <T: Any> getFromPlatform(clazz: KClass<T>,buffer: Buffer): T
